Add custom hooks folder , Refactor theme preference handling and chat toggle
