+++ frontend/src/hooks/useThemePreference.js
@@ -1,0 +1,10 @@
+import { useEffect } from 'react';
+
+export function useThemePreference(setDarkMode){
+    useEffect(() => {
+      const savedDarkMode = localStorage.getItem('darkMode');
+      if (savedDarkMode) {
+        setDarkMode(savedDarkMode === 'true');
+      }
+    }, [setDarkMode]);
+};
